+++ tests/integration/cases/empty_dict/Rust_declaration_style_lazy_static.rs
@@ -1,0 +1,6 @@
+use std::sync::LazyLock;
+use std::collections::HashMap;
+fn main() {
+    static my_data: LazyLock<HashMap<String, String>> = LazyLock::new(|| HashMap::<String, String>::from([]));
+    let _ = my_data;
+}