+++ tests/integration/cases/empty_list/Rust_declaration_style_lazy_static.rs
@@ -1,0 +1,5 @@
+use std::sync::LazyLock;
+fn main() {
+    static my_data: LazyLock<Vec<String>> = LazyLock::new(|| Vec::<String>::new());
+    let _ = my_data;
+}