+++ tests/integration/cases/int_key_dict/Nim_declaration_style_let.nim
@@ -1,0 +1,6 @@
+import json
+let my_data = %* {
+    "1": "one",
+    "2": "two",
+    "42": "answer"
+}