+++ tests/integration/cases/int_list/Nim_declaration_style_let.nim
@@ -1,0 +1,6 @@
+import json
+let my_data = @[
+    1,
+    2,
+    3
+]