+++ tests/integration/cases/int_list/Rust_declaration_style_const.rs
@@ -1,0 +1,8 @@
+fn main() {
+    const my_data: [i32; 3] = [
+        1,
+        2,
+        3,
+    ];
+    let _ = my_data;
+}