+++ tests/integration/cases/int_list/V_declaration_style_mut.v
@@ -1,0 +1,9 @@
+
+fn main() {
+	mut my_data := [
+		1,
+		2,
+		3,
+	]
+	_ = my_data
+}